+++ packages/components/badge/src/badge.stories.ts
@@ -1,9 +1,9 @@
-import type { Badge, BadgeSize, BadgeVariant } from './badge.js';
-import type { Meta, StoryObj } from '@storybook/web-components';
-import { Icon } from '@sl-design-system/icon';
 import { faCheck, faGear } from '@fortawesome/pro-regular-svg-icons';
+import { Icon } from '@sl-design-system/icon';
+import { type Meta, type StoryObj } from '@storybook/web-components';
 import { html, nothing } from 'lit';
 import '../register.js';
+import { type Badge, type BadgeSize, type BadgeVariant } from './badge.js';
 
 type Props = Pick<Badge, 'size' | 'variant'> & { icon?: boolean; text?: string };
 
@@ -51,12 +51,11 @@ export default {
       options: variants
     }
   },
-  render: ({ size, text, variant, icon }) =>
-    html`
-      <sl-badge .size=${size} .variant=${variant}
-        >${icon ? html`<sl-icon name="check"></sl-icon>` : nothing}${text}</sl-badge
-      >
-    `
+  render: ({ size, text, variant, icon }) => html`
+    <sl-badge .size=${size} .variant=${variant}>
+      ${icon ? html`<sl-icon name="check"></sl-icon>` : nothing} ${text}
+    </sl-badge>
+  `
 } satisfies Meta<Props>;
 
 export const Basic: Story = {};
@@ -89,16 +88,18 @@ export const All: Story = {
         </thead>
         <tbody>
           ${sizes.map(
-            size => html` <tr>
-              <th>${sizeName(size)}</th>
-              ${variants.map(
-                variant => html`<td>
-                  <sl-badge .variant=${variant} .size=${size}><sl-icon name="far-check"></sl-icon></sl-badge>
-                  <sl-badge .variant=${variant} .size=${size}>99+</sl-badge>
-                  <sl-badge .variant=${variant} .size=${size}><sl-icon name="far-gear"></sl-icon> away</sl-badge>
-                </td>`
-              )}
-            </tr>`
+            size =>
+              html` <tr>
+                <th>${sizeName(size)}</th>
+                ${variants.map(
+                  variant =>
+                    html`<td>
+                      <sl-badge .variant=${variant} .size=${size}><sl-icon name="far-check"></sl-icon></sl-badge>
+                      <sl-badge .variant=${variant} .size=${size}>99+</sl-badge>
+                      <sl-badge .variant=${variant} .size=${size}><sl-icon name="far-gear"></sl-icon> away</sl-badge>
+                    </td>`
+                )}
+              </tr>`
           )}
         </tbody>
       </table>