+++ packages/components/form/src/form-field.stories.ts
@@ -1,15 +1,15 @@
-import type { FormField } from './form-field.js';
-import type { Meta, StoryObj } from '@storybook/web-components';
 import '@sl-design-system/checkbox/register.js';
 import '@sl-design-system/icon/register.js';
 import '@sl-design-system/radio-group/register.js';
 import '@sl-design-system/select/register.js';
 import '@sl-design-system/switch/register.js';
-import '@sl-design-system/textarea/register.js';
 import '@sl-design-system/text-field/register.js';
+import '@sl-design-system/textarea/register.js';
 import '@sl-design-system/tooltip/register.js';
+import { type Meta, type StoryObj } from '@storybook/web-components';
 import { type TemplateResult, html } from 'lit';
 import '../register.js';
+import { type FormField } from './form-field.js';
 
 type Props = Pick<FormField, 'hint' | 'label'> & {
   customValidity?: string;