+++ packages/components/form/src/label.stories.ts
@@ -1,11 +1,11 @@
-import type { StoryObj } from '@storybook/web-components';
 import '@sl-design-system/button/register.js';
 import '@sl-design-system/button-bar/register.js';
 import '@sl-design-system/checkbox/register.js';
 import '@sl-design-system/icon/register.js';
+import '@sl-design-system/switch/register.js';
 import '@sl-design-system/text-field/register.js';
 import '@sl-design-system/tooltip/register.js';
-import '@sl-design-system/switch/register.js';
+import { type StoryObj } from '@storybook/web-components';
 import { html } from 'lit';
 import '../register.js';