+++ packages/components/skeleton/src/skeleton.ts
@@ -1,5 +1,4 @@
-import type { CSSResultGroup } from 'lit';
-import { LitElement } from 'lit';
+import { type CSSResultGroup, LitElement } from 'lit';
 import { property } from 'lit/decorators.js';
 import styles from './skeleton.scss.js';
 
@@ -11,18 +10,20 @@ export class Skeleton extends LitElement {
   /** @private */
   static override styles: CSSResultGroup = styles;
 
-  /** Skeleton's effect.
-   *  @type {'none' | 'shimmer' | 'pulse' | 'sheen'}
-   * */
+  /**
+   * Skeleton's effect.
+   * @type {'none' | 'shimmer' | 'pulse' | 'sheen'}
+   */
   @property({ reflect: true }) effect: SkeletonEffect = 'shimmer';
 
-  /** Skeleton's variant.
-   *  @type {'circle' | 'default'}
-   * */
+  /**
+   * Skeleton's variant.
+   * @type {'circle' | 'default'}
+   */
   @property({ reflect: true }) variant: SkeletonVariant = 'default';
 
-  constructor() {
-    super();
+  override connectedCallback(): void {
+    super.connectedCallback();
 
     this.setAttribute('aria-busy', 'true');
   }